Added Sample - Item DATA JSON
p.s.

personal thought

Item Hash - basic archetecture for Weapon
Item Instance - Check the perks assigned to that weapon instance

Both needed to be combined to calculate the final stat of the weapon.

With ItemHash, we can still get base stat, but stat changes based on the perks, and not all the perks are abailable at the same time. so that is the reason why we need to combine both, I believe.